+++ src/app/components/model_tree/model_tree_line.html
@@ -14,12 +14,28 @@
         </span>
     </a>
 
-    <a ng-if="item.unique_id"
+    <a ng-if="item.unique_id && item.node_type == 'model'"
        class='unselectable'
        ng-class="getClass(item)"
        ng-click="activate(item)"
        data-nav-unique-id="{{ item.unique_id }}"
        ui-sref='dbt.model({unique_id: item.unique_id})'>
+        <span class="filename">
+            <span class="filename-normal">
+                <svg class="icn menu-icon-on"><use xlink:href="{{ getIcon(item.type, 'on') }}"></use></svg>
+                <svg class="icn menu-icon-off"><use xlink:href="{{ getIcon(item.type, 'off') }}"></use></svg>
+            </span>
+            <span class="filename-ellip">{{name.start}}</span>
+            <span class="filename-normal">{{name.end}}</span>
+        </span>
+    </a>
+
+    <a ng-if="item.unique_id && item.node_type == 'source'"
+       class='unselectable'
+       ng-class="getClass(item)"
+       ng-click="activate(item)"
+       data-nav-unique-id="{{ item.unique_id }}"
+       ui-sref='dbt.source({unique_id: item.unique_id})'>
         <span class="filename">
             <span class="filename-normal">
                 <svg class="icn menu-icon-on"><use xlink:href="{{ getIcon(item.type, 'on') }}"></use></svg>
