+++ src/app/index.routes.js
@@ -6,6 +6,7 @@ const templates = {
     overview: require('./overview/overview.html'),
     graph: require('./graph/graph.html'),
     model: require('./docs/model.html'),
+    source: require('./docs/source.html'),
 }
 
 angular
@@ -36,6 +37,14 @@ angular
             url: 'model/:unique_id?section&' + graph_params,
             controller: 'ModelCtrl',
             templateUrl: templates.model,
+            params: {
+                unique_id: {type: 'string'}
+            },
+        })
+        .state('dbt.source', {
+            url: 'source/:unique_id?section&' + graph_params,
+            controller: 'SourceCtrl',
+            templateUrl: templates.source,
             params: {
                 unique_id: {type: 'string'}
             },
